feat: added user personal data update

+++ authentication_service/infrastructure/api/pb_mapper.go
@@ -7,7 +7,7 @@ import (
 
 func mapUser(user *domain.User) *pb.User {
 	userPb := &pb.User{
-		Id:       user.ID,
+		Id:       int64(user.ID),
 		Name:     user.Name,
 		Username: user.Username,
 		Password: user.Password,
